Restructure, migrate create.py
